refactor: move IntegrateJordan and IntegrateShape to geometry.integral
