+++ src/shapepy/geometry/integral.py
@@ -4,110 +4,101 @@ Contains functions to integrate over a segment
 
 from __future__ import annotations
 
-from typing import Union
+from typing import Optional, Union
 
 import numpy as np
 
-from ..scalar.quadrature import (
-    AdaptativeIntegrator,
-    IntegratorFactory,
-    closed_linspace,
-)
+from ..scalar.bezier import Bezier, bezier2polynomial
+from ..scalar.polynomial import derivate, integrate
+from ..scalar.quadrature import closed_linspace
 from ..scalar.reals import Math
 from ..tools import Is
+from .jordancurve import JordanCurve
 from .point import Point2D
 from .segment import Segment
 
 
-def vertical(
-    curve: Segment,
-    expx: int = 0,
-    expy: int = 0,
-    nnodes: Union[None, int] = None,
-):
-    """Computes the integral I
-
-    I = int_C x^expx * y^expy * dy
-
+class IntegrateSegment:
     """
-    if not Is.instance(curve, Segment):
-        raise TypeError
-    if not Is.integer(expx) or expx < 0:
-        raise ValueError
-    if not Is.integer(expy) or expy < 0:
-        raise ValueError
-    if nnodes is None:
-        nnodes = 3 + expx + expy + curve.degree
-    elif not Is.integer(nnodes) or nnodes < 0:
-        raise ValueError
-    dcurve = curve.derivate()
-    direct = IntegratorFactory.custom_open_formula(nnodes)
-
-    def function(node):
-        xval, yval = curve(node)
-        _, dyval = dcurve(node)
-        return dyval * (xval**expx) * (yval**expy)
-
-    return direct.integrate(function, (0, 1))
-
-
-def polynomial(
-    curve: Segment, expx: int, expy: int, nnodes: Union[None, int] = None
-):
-    """
-    Computes the integral
-
-    I = int_C x^expx * y^expy * ds
-    """
-    if not Is.instance(curve, Segment):
-        raise TypeError
-    if nnodes is None:
-        nnodes = 3 + expx + expy + curve.degree
-    elif not Is.integer(nnodes) or nnodes < 0:
-        raise ValueError
-    assert expx == 0
-    assert expy == 0
-
-    direct = IntegratorFactory.open_newton_cotes(nnodes)
-
-    def function(node):
-        xval, yval = curve(node)
-        return xval**expx * yval**expy
-
-    return direct.integrate(function, (0, 1))
-
-
-def lenght(curve: Segment):
-    """Computes the integral I
-
-        I = int_{C} ds
-
-    Given the control points P of a bezier curve C(u) of
-    degree p
-
-        C(u) = sum_{i=0}^{p} B_{i,p}(u) * P_i
-
-        I = int_{0}^{1} abs(C'(u)) * du
-
+    Defines methods to integrates over a segment
     """
 
-    dcurve = curve.derivate()
-    direct = IntegratorFactory.clenshaw_curtis(3)
-    adaptive = AdaptativeIntegrator(direct, 1e-9, 5)
+    @staticmethod
+    def polynomial(curve: Segment, expx: int, expy: int):
+        """
+        Computes the integral
 
-    def function(node):
-        return abs(dcurve(node))
+        I = int_D x^expx * y^expy * dA
 
-    return adaptive.integrate(function, (0, 1))
+        """
+        if not Is.instance(curve, Segment):
+            raise TypeError
+
+        xfunc = bezier2polynomial(Bezier(pt[0] for pt in curve.ctrlpoints))
+        yfunc = bezier2polynomial(Bezier(pt[1] for pt in curve.ctrlpoints))
+
+        poly = (xfunc**expx) * (yfunc**expy)
+        poly *= xfunc * derivate(yfunc) - yfunc * derivate(xfunc)
+        poly = integrate(poly)
+        return (poly(1) - poly(0)) / (expx + expy + 2)
+
+    @staticmethod
+    def winding_number(
+        curve: Segment,
+        center: Point2D = (0.0, 0.0),
+        nnodes: Union[None, int] = None,
+    ) -> float:
+        """
+        Computes the integral for a bezier curve of given control points
+        """
+        assert Is.instance(curve, Segment)
+        nnodes = curve.npts if nnodes is None else nnodes
+        nodes = closed_linspace(nnodes)
+        total = 0
+        for nodea, nodeb in zip(nodes[:-1], nodes[1:]):
+            pointa = curve(nodea)
+            pointb = curve(nodeb)
+            total += winding_number_linear(pointa, pointb, center)
+        return total
 
 
-def area(curve: Segment, nnodes: Union[None, int] = None):
-    """Computes the integral I
-
-    I = int_0^1 x * dy
-
+class IntegrateJordan:
     """
-    return vertical(curve, 1, 0, nnodes)
+    Defines functions to integrate over the internal area
+    defined by the jordan curve.
+    """
+
+    @staticmethod
+    def polynomial(jordan: JordanCurve, expx: int, expy: int):
+        """
+        Computes the integral
+
+        I = int x^expx * y^expy * ds
+        """
+        assert Is.instance(jordan, JordanCurve)
+        return sum(
+            IntegrateSegment.polynomial(segment, expx, expy)
+            for segment in jordan.segments
+        )
+
+    @staticmethod
+    def winding_number(
+        jordan: JordanCurve,
+        center: Optional[Point2D] = (0.0, 0.0),
+        nnodes: Optional[int] = None,
+    ) -> Union[int, float]:
+        """Computes the winding number from jordan curve
+
+        Returns [-1, -0.5, 0, 0.5 or 1]
+        """
+        wind = 0
+        if center in jordan.box():
+            for bezier in jordan.segments:
+                if center in bezier:
+                    return 0.5 if float(jordan) > 0 else -0.5
+        for bezier in jordan.segments:
+            wind += IntegrateSegment.winding_number(bezier, center, nnodes)
+        return round(wind)
 
 
 def winding_number_linear(
@@ -128,22 +119,3 @@ def winding_number_linear(
     if abs(wind) < 0.5:
         return wind
     return wind - 1 if wind > 0 else wind + 1
-
-
-def winding_number(
-    curve: Segment,
-    center: Point2D = (0.0, 0.0),
-    nnodes: Union[None, int] = None,
-) -> float:
-    """
-    Computes the integral for a bezier curve of given control points
-    """
-    assert Is.instance(curve, Segment)
-    nnodes = curve.npts if nnodes is None else nnodes
-    nodes = closed_linspace(nnodes)
-    total = 0
-    for nodea, nodeb in zip(nodes[:-1], nodes[1:]):
-        pointa = curve(nodea)
-        pointb = curve(nodeb)
-        total += winding_number_linear(pointa, pointb, center)
-    return total
